pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enterprise-ai-platform-sdk"
version = "0.3.0"
description = "Enterprise AI Platform — shared SDK for agents, MCP servers, security, caching, and compaction"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "opentelemetry-instrumentation-openai>=0.25.0",
    "structlog>=24.1.0",
    # Observability — LangFuse for LLM observability and prompt management
    "langfuse>=2.0.0,<3.0.0",
    # Security — OPA client and API key verifier
    "httpx>=0.27.0",
    # Caching — Redis-backed tool result cache
    "redis[asyncio]>=5.0.0",
    # Compaction — accurate token counting
    "tiktoken>=0.7.0",
    # Agent factory (langchain required by langfuse CallbackHandler integration)
    "langchain>=0.2.0,<1.0.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    # Structured output (RMBrief, specialist synthesis)
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.111.0"]
mcp-bridge = ["mcp>=1.0.0"]
postgres-checkpointer = ["langgraph-checkpoint-postgres>=2.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["platform_sdk*"]

[tool.ruff]
line-length = 100
target-version = "py311"
